requirements.txt
PySide6>=6.5.0
PyInstaller>=5.0.0
Pillow>=9.0.0
qtawesome>=1.2.3
psutil>=5.9.0

